adding function to generate random quiz_id

+++ db/schema/03_qandas.sql
@@ -1,14 +1,14 @@
 DROP TABLE IF EXISTS question CASCADE;
 CREATE TABLE question (
   id SERIAL PRIMARY KEY NOT NULL,
-  quiz_id INTEGER REFERENCES quiz(id),
+  quiz_id BIGINT REFERENCES quiz(id),
   content TEXT
 );
 
 DROP TABLE IF EXISTS answer CASCADE;
 CREATE TABLE answer (
   id SERIAL PRIMARY KEY NOT NULL,
-  question_id INTEGER REFERENCES question(id),
+  question_id BIGINT REFERENCES question(id),
   content TEXT,
   correct BOOLEAN
 );
