+++ db/schema/04_attempts.sql
@@ -2,7 +2,7 @@ DROP TABLE IF EXISTS attempt CASCADE;
 CREATE TABLE attempt (
   id SERIAL PRIMARY KEY NOT NULL,
   user_id INTEGER REFERENCES users(id),
-  quiz_id INTEGER REFERENCES quiz(id),
+  quiz_id BIGINT REFERENCES quiz(id),
   correct_responses INTEGER,
   total_responses INTEGER
 );
